requirements.txt
pipecat-ai[runner,soniox,openai,webrtc]
python-dotenv
loguru
torch
certifi